Moved the MixSeeker and MixSeeker test to the new seeker package.
