+++ app/models/seeker/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * This package contains all of the classes that seek the starttime of a
+ * snippet.
+ * 
+ * @since 26-05-2015
+ */
+package models.seeker;
